Simplify
This removes stats and archetypes. While I liked the idea, I decided
that the complex UI necessary for these features is out of scope for the
initial version of this project. Moving forward, roll creation will be
handled by parsing CSV data (that was likely generated outside of the
app).